+++ apps/web/next.config.js
@@ -9,6 +9,14 @@ const nextConfig = {
   images: {
     domains: ['via.placeholder.com', 'source.unsplash.com'],
   },
+  webpack(config) {
+    config.module.rules.push({
+      test: /\.svg$/,
+      use: ['@svgr/webpack'],
+    })
+
+    return config
+  },
 }
 
 module.exports = shouldAnalyzeBundle