+++ apps/web/src/app/layoutConfig.tsx
@@ -1,33 +1,61 @@
 import React from 'react'
 import Image from 'next/image'
-import { Link, useSettings } from '@gravis-os/ui'
+import { Stack, Typography, Link, useSettings } from '@gravis-os/ui'
+import LaunchOutlinedIcon from '@mui/icons-material/LaunchOutlined'
+import EmblemSvg from '../../public/emblem.svg'
 
 const logoJsx = (
   <Link href="/" sx={{ display: 'flex' }}>
-    <Image src="/logo.png" width={40} height={40} />
+    <Image src="/emblem_dark.svg" width={40} height={40} />
+  </Link>
+)
+
+// TODO@Joel: Abstract this
+const logoSvg = (
+  <Link
+    href="/"
+    underline="none"
+    sx={{
+      display: 'flex',
+      '& svg': {
+        width: 20,
+        fill: ({ palette }) => palette.text.primary,
+      },
+    }}
+  >
+    <Stack direction="row" alignItems="center" spacing={1.5}>
+      <EmblemSvg />
+      <Typography variant="overline">Gravis OS</Typography>
+    </Stack>
   </Link>
 )
 
 const landingLayoutProps = {
   headerProps: {
+    color: 'transparent',
+    disableBoxShadow: true,
     navItems: {
       left: [
         {
           key: 'logo',
           title: 'Logo',
-          children: logoJsx,
+          children: logoSvg,
           sx: { mr: 1 },
           showOnMobileBar: true,
         },
+      ],
+      right: [
         {
           key: 'docs',
           title: 'Docs',
           href: 'https://docs.gravis-os.com',
           linkProps: { target: '_blank' },
+          // TODO@Joel: Abstract this to linkProps by detecting the target blank
+          buttonProps: {
+            endIcon: <LaunchOutlinedIcon />,
+            size: 'small',
+          },
         },
-        { key: 'features', title: 'Features', href: '/features' },
-      ],
-      right: [
         {
           key: 'dark-mode',
           title: 'Dark Mode',