+++ packages/eslint-config/.eslintrc.js
@@ -10,6 +10,7 @@ module.exports = {
   env: {
     browser: true,
     es6: true,
+    jest: true,
   },
   globals: {
     Atomics: 'readonly',
@@ -22,7 +23,14 @@ module.exports = {
     'plugin:@typescript-eslint/recommended',
     'prettier',
   ],
-  plugins: ['react', 'import', 'prettier', '@typescript-eslint'],
+  plugins: [
+    'react',
+    'import',
+    'prettier',
+    '@typescript-eslint',
+    'jest',
+    'testing-library',
+  ],
   rules: {
     // Javascript
     'arrow-body-style': 0,