+++ packages/form/src/form/fields/DateField.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { DatePicker, DatePickerProps } from '@mui/x-date-pickers'
 import { TextField, TextFieldProps } from '@mui/material'
 
-interface DateFieldProps extends DatePickerProps<Date> {
+interface DateFieldProps extends DatePickerProps<Date, any> {
   textFieldProps?: Partial<Omit<TextFieldProps, 'variant'>>
 }