+++ packages/web/src/themes/Landing/gravisOsLandingTheme.ts
@@ -1,0 +1,67 @@
+import merge from 'lodash/merge'
+import { baseTheme } from '@gravis-os/ui'
+import landingTheme from './landingTheme'
+
+const headerFontFamily =
+  "SF Pro Display,-apple-system, Roboto, 'Helvetica Neue', Arial, sans-serif"
+const subtitleFontFamily =
+  "SF Pro Text,-apple-system, Roboto, 'Helvetica Neue', Arial, sans-serif"
+
+const gravisLandingOsThemeConfig = {
+  typography: {
+    h1: {
+      fontFamily: headerFontFamily,
+      fontSize: '4rem',
+      [baseTheme.light.breakpoints.up('sm')]: { fontSize: '4.25rem' },
+      [baseTheme.light.breakpoints.up('md')]: { fontSize: '5.5rem' },
+      [baseTheme.light.breakpoints.up('lg')]: { fontSize: '6rem' },
+      lineHeight: 1.05,
+      fontWeight: 'bold',
+      letterSpacing: '-0.02em',
+      '&.MuiTypography-gutterBottom': { marginBottom: '0.2em' },
+    },
+    h2: {
+      fontFamily: headerFontFamily,
+      lineHeight: 1.15555556,
+      letterSpacing: '-0.025em',
+      fontSize: '3rem',
+      [baseTheme.light.breakpoints.up('sm')]: { fontSize: '3.5rem' },
+      [baseTheme.light.breakpoints.up('md')]: { fontSize: '3.75rem' },
+      [baseTheme.light.breakpoints.up('lg')]: { fontSize: '4rem' },
+    },
+    h3: {
+      fontFamily: headerFontFamily,
+      fontSize: '2rem',
+      [baseTheme.light.breakpoints.up('sm')]: { fontSize: '2.25rem' },
+      [baseTheme.light.breakpoints.up('md')]: { fontSize: '2.5rem' },
+      [baseTheme.light.breakpoints.up('lg')]: { fontSize: '2.75rem' },
+      fontWeight: 600,
+    },
+    h4: { fontFamily: headerFontFamily },
+    h5: { fontFamily: subtitleFontFamily },
+    subtitle2: {
+      fontFamily: subtitleFontFamily,
+      fontSize: '1rem',
+      fontWeight: 'normal',
+      [baseTheme.light.breakpoints.up('sm')]: { fontSize: '1rem' },
+      [baseTheme.light.breakpoints.up('md')]: { fontSize: '1rem' },
+      [baseTheme.light.breakpoints.up('lg')]: { fontSize: '1.25rem' },
+      lineHeight: 1.33,
+    },
+    body1: {
+      fontFamily: subtitleFontFamily,
+      fontSize: '1rem',
+      [baseTheme.light.breakpoints.up('sm')]: { fontSize: '1rem' },
+      [baseTheme.light.breakpoints.up('md')]: { fontSize: '1rem' },
+      [baseTheme.light.breakpoints.up('lg')]: { fontSize: '1.25rem' },
+      lineHeight: 1.33,
+    },
+  },
+}
+
+const gravisOsLandingTheme = {
+  light: merge({}, landingTheme.light, gravisLandingOsThemeConfig),
+  dark: merge({}, landingTheme.dark, gravisLandingOsThemeConfig),
+}
+
+export default gravisOsLandingTheme