+++ apps/web/CHANGELOG.md
@@ -1,5 +1,74 @@
 # web
 
+## 0.0.11
+
+### Patch Changes
+
+- [`d518d51`](https://github.com/gravis-os/gravis-os/commit/d518d51cfa05e35d0c6d83a6bb39f76955e77611) Thanks [@clodal](https://github.com/clodal)! - Fix deployment issue with apps/web
+
+- Updated dependencies [[`d518d51`](https://github.com/gravis-os/gravis-os/commit/d518d51cfa05e35d0c6d83a6bb39f76955e77611)]:
+  - @gravis-os/form@0.0.55
+  - @gravis-os/ui@0.0.28
+  - @gravis-os/web@0.0.15
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`a9b4f8b`](https://github.com/gravis-os/gravis-os/commit/a9b4f8b07dad99d671fb54d697aaa5fedcbe099a)]:
+  - @gravis-os/form@0.0.53
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`bbfabaf`](https://github.com/gravis-os/gravis-os/commit/bbfabaf5d155b4c1fc29c2b2d0d0103c735703e2)]:
+  - @gravis-os/form@0.0.52
+
+## 0.0.8
+
+### Patch Changes
+
+- [`cad87f1`](https://github.com/gravis-os/gravis-os/commit/cad87f1add9c118193ac6adf4dcfe75cadca3260) Thanks [@clodal](https://github.com/clodal)! - Update web landing page
+
+- Updated dependencies [[`cad87f1`](https://github.com/gravis-os/gravis-os/commit/cad87f1add9c118193ac6adf4dcfe75cadca3260)]:
+  - @gravis-os/form@0.0.51
+  - @gravis-os/ui@0.0.27
+  - @gravis-os/web@0.0.14
+
+## 0.0.7
+
+### Patch Changes
+
+- Update Block padding config
+
+- Updated dependencies [[`cfc0c4f`](https://github.com/gravis-os/gravis-os/commit/cfc0c4f161a985e015cc17d6e8f08d8d440a4185)]:
+  - @gravis-os/web@0.0.13
+  - @gravis-os/form@0.0.50
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`84290d9`](https://github.com/gravis-os/gravis-os/commit/84290d99d1cdf998292b6e6525abc14f693a350b)]:
+  - @gravis-os/form@0.0.49
+
+## 0.0.5
+
+### Patch Changes
+
+- [`411605d`](https://github.com/gravis-os/gravis-os/commit/411605da5889c6d66936ae27d73a1d647c53b5b7) Thanks [@clodal](https://github.com/clodal)! - Add toggleDarkModeIconButtonJsx to SettingsProvider
+
+* [`2646139`](https://github.com/gravis-os/gravis-os/commit/2646139864af3c62032571e6f17383505a1298da) Thanks [@clodal](https://github.com/clodal)! - Update web to recursively render stackItems"
+
+- [`343f0f8`](https://github.com/gravis-os/gravis-os/commit/343f0f8cfd9f98eea1ef7ad220f6e66aa886a536) Thanks [@clodal](https://github.com/clodal)! - Update web UI styles"
+
+* [`073a373`](https://github.com/gravis-os/gravis-os/commit/073a373a84e5c92acc6d8b511c883775910a44a3) Thanks [@clodal](https://github.com/clodal)! - Add Typography.gradient"
+
+* Updated dependencies [[`411605d`](https://github.com/gravis-os/gravis-os/commit/411605da5889c6d66936ae27d73a1d647c53b5b7), [`feb25a5`](https://github.com/gravis-os/gravis-os/commit/feb25a5bc5c9292f32657e3bbaf2d860d37b0c48), [`2646139`](https://github.com/gravis-os/gravis-os/commit/2646139864af3c62032571e6f17383505a1298da), [`c956526`](https://github.com/gravis-os/gravis-os/commit/c9565260b8b10ef1834613bd045c89b90a482a32), [`343f0f8`](https://github.com/gravis-os/gravis-os/commit/343f0f8cfd9f98eea1ef7ad220f6e66aa886a536), [`8115921`](https://github.com/gravis-os/gravis-os/commit/811592173c598632d6eba945502d1cc0906c734b), [`073a373`](https://github.com/gravis-os/gravis-os/commit/073a373a84e5c92acc6d8b511c883775910a44a3), [`d5f769b`](https://github.com/gravis-os/gravis-os/commit/d5f769be4fd88388f730a1fe50ac5927d32a61fc), [`d5721b8`](https://github.com/gravis-os/gravis-os/commit/d5721b89df7c5c7650f3e916e23fd9cc141ea8e1), [`cbe728e`](https://github.com/gravis-os/gravis-os/commit/cbe728e2daa70af1c2fb1406cc1691d302518110), [`b96d5c1`](https://github.com/gravis-os/gravis-os/commit/b96d5c163a49fc77f11f8c97da72c6ab2819ed92)]:
+  - @gravis-os/ui@0.0.26
+  - @gravis-os/web@0.0.12
+
 ## 0.0.4
 
 ### Patch Changes