+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gravis-os/tsconfig
 
+## 0.0.5
+
+### Patch Changes
+
+- [`6c0d753`](https://github.com/gravis-os/gravis-os/commit/6c0d75395e7d62966fba1f6a95c36c7ed707feba) Thanks [@clodal](https://github.com/clodal)! - Fix build error in web app
+
+## 0.0.4
+
+### Patch Changes
+
+- [`a9b4f8b`](https://github.com/gravis-os/gravis-os/commit/a9b4f8b07dad99d671fb54d697aaa5fedcbe099a) Thanks [@clodal](https://github.com/clodal)! - Update nextjs tsconfig to fix build errors on Vercell
+
+## 0.0.3
+
+### Patch Changes
+
+- [`bbfabaf`](https://github.com/gravis-os/gravis-os/commit/bbfabaf5d155b4c1fc29c2b2d0d0103c735703e2) Thanks [@clodal](https://github.com/clodal)! - Update tsconfig to disable implicit any for nextjs build to pass
+
 ## 0.0.2
 
 ### Patch Changes